Create separate module for intrinsic typechecking.
